refactor: split core/models.py into 6-module package
- Created core/models/ package with domain-focused modules:
  - property.py (Property, RentalIncome, InvestmentAnalysis, etc.)
  - pipeline.py (PipelineProperty, PipelineAsset, ScreeningCriteria, etc.)
  - sources.py (VrmProperty, HudProperty, UsdaProperty, etc.)
  - growth.py (GrowthArea, MarketSnapshot, Listing, etc.)
  - notifications.py (Notification, UserWatchlist, etc.)
  - base.py (Team, TeamMember, PropertyNote, UserProfile, etc.)
- __init__.py re-exports all models (backward compatible)
- Added cross-module imports for ForeignKey dependencies
- Updated import order to avoid circular dependencies
- All 42 models pass system check and tests
